+++ R/MetricGraph-package.R
@@ -33,10 +33,11 @@
 #' @import Matrix
 #' @export Cholesky t
 #' @importFrom igraph make_graph distances is_tree
-#' @importFrom rSPDE matern.covariance gg_df create_train_test_indices
+#' @importFrom rSPDE matern.covariance gg_df create_train_test_indices posterior_crossvalidation
 #' @importFrom rlang .data
 #' @importFrom methods is slot new as
 #' @export gg_df
+#' @export posterior_crossvalidation
 #' @importFrom stats predict approx nobs deviance lm logLik na.omit dist sd dnorm pnorm rnorm var as.formula delete.response model.matrix optim terms simulate rpois runif profile qchisq spline
 #' @importFrom ggplot2 ggplot geom_path aes geom_point coord_fixed labs scale_colour_gradientn guide_legend scale_colour_discrete
 #' @importFrom spatstat.geom psp owin